requirements.txt

reflex==0.8.25
supabase==2.10.0
python-dotenv==1.0.0
cryptography==42.0.0
web3==6.15.1
eth-abi==5.0.0
hyperliquid==0.4.66
hyperliquid-python-sdk==0.20.1